add placeholder warning

+++ clear.py
@@ -2,11 +2,13 @@
 
 import os
 
-for filename in os.scandir():
-    if filename.is_file() and filename.path != '.\clear.py':
-        print (filename.path)
-        with open(filename.path,'w') as file:
-            file.write('''{
+warning = input('WARNING: running this script will overwrite ALL files in this directory and corresponding directories. are you sure? [y/n] ')
+if warning == 'y':
+    for filename in os.scandir():
+        if filename.is_file() and filename.path != '.\clear.py':
+            print (filename.path)
+            with open(filename.path,'w') as file:
+                file.write('''{
     "type": "minecraft:crafting_shaped",
     "key": {
         "#": {
@@ -21,4 +23,6 @@ for filename in os.scandir():
     }
 }''')
 
-print ('finished')
+    print ('finished')
+else:
+    print ('aborted')
